fix: move module to a jsDelivr directory
